Create README.md
About
This is a Node.js command-line application that allows users to enter a country name via the terminal. The application uses the Mapbox API to retrieve the location's coordinates (latitude and longitude) and the WeatherAPI to fetch the current weather conditions and temperature. The results are then displayed in the console